Edit function

+++ Crud1/Models/Product.cs
@@ -6,7 +6,7 @@ namespace Crud1.Models
     public class Product
     {
         [Key]
-        public double Pro_id {  get; set; }
+        public int Pro_id {  get; set; }
         public string pro_name { get; set; }
         public double pro_price { get; set; }
         public string pro_description { get; set; }
